i2c: SCL pin should be both output and input
ESP-IDF itself will set both pins as INPUT_OUTPUT_OD, so we should
ensure that the pins are capable of that.

The reason behind that is probably because both the master and slave
actually need to be able to read and write the clock line.  I2C has a
thing called "clock stretching" where a slave device will hold the clock
line low to indicate that it needs a little time to complete the
master's request.  So in that case the slave needs output as well as
input, and the master needs input as well as output.