Generate bindings for extra components
There are two ways the bindings are generated.

The first is when `bindings_module` is not set. In that case the header
of the extra component will just be added to the normal esp-idf bindings
generation. All additional bindings will also be in the root of `esp-idf-sys`.

The second: when `bindings_module` is set. For every unique `bindings_module`
bindgen will be invoked to generate the bindings, which will then be appended as a new
module to the `bindings.rs` file. Note though that since this is a new bindgen invocation
duplicate symbols of the normal esp-idf bindings may be generated.