Fix tools install dir not getting applied correctly
When an activated esp-idf environment was detected while `esp_idf_tools_install_dir` was unset
instead of forwarding the `ESP_IDF_TOOLS_DIR` env var to the cmake build, the default
tools install dir (in the workspace directory) was forwarded instead.
This caused the esp-idf build system to use a non-existent python in that directory and failed to build.

Addtionally, fixes the optional cargo metadata setting `idf_path` being ignored. And the global tools install
directory of the `embuild::espidf::Installer` is now explicitly forwarded to cmake instead of relying on
esp-idf defaults.