build: avoid using custom cmake on esp-idf 5.x (where x < 4) (#191)
Fixes the version check for when to install a custom cmake.

The check is supposed to make sure that all versions before `v4.4` get a more 
recent cmake by supplying a custom `tools.json`. Instead of this, the custom 
`tools.json` was used for *all* `version<x.4` (where `x` is some major version).

This would not be a problem in itself, but unfortunately the custom `tools.json`
doesn't provide a binary for the `macos-arm64` platform, thus resulting in a
failed build.