PCNT v5 bindings take 2 (#230)
* Provide PCNT v5 bindings.

This provides bindings for the PCNT peripheral using the ESP-IDF v5+
definitions in conjunction with the ESP-IDF v4 definitions. It removes
a case where bindgen can't cope with a typename that exists in two
namespaces (enum pcnt_unit_t in v4, struct pcnt_unit_t in v5) by
manually handling this special case.

* Fix PCNT bindings.

This fixes the PCNT bindings so that they work with esp-idf-hal,
and actually tested against chips that *have* PCNT peripherals this
time.