spi: Add custom spi::Operation enum
This replaces the Operation enum which was previously re-exported from
embedded_hal.

As of this commit, the enum is identical to the embedded_hal Operation,
except that there is no Word type parameter, as esp-idf-hal only uses u8
for SPI. Additionally, the `#[non_exhaustive]` attribute has been added
to allow for future expansion.